Move util files
