+++ site/src/components/WorkspaceScheduleForm/WorkspaceScheduleForm.stories.tsx
@@ -3,8 +3,8 @@ import dayjs from "dayjs"
 import advancedFormat from "dayjs/plugin/advancedFormat"
 import timezone from "dayjs/plugin/timezone"
 import utc from "dayjs/plugin/utc"
-import { defaultSchedule, emptySchedule } from "pages/WorkspacesPage/schedule"
-import { emptyTTL } from "pages/WorkspacesPage/ttl"
+import { defaultSchedule, emptySchedule } from "pages/WorkspaceSchedulePage/schedule"
+import { emptyTTL } from "pages/WorkspaceSchedulePage/ttl"
 import { makeMockApiError } from "testHelpers/entities"
 import { WorkspaceScheduleForm, WorkspaceScheduleFormProps } from "./WorkspaceScheduleForm"
