+++ site/src/components/WorkspaceScheduleForm/WorkspaceScheduleForm.tsx
@@ -17,8 +17,8 @@ import relativeTime from "dayjs/plugin/relativeTime"
 import timezone from "dayjs/plugin/timezone"
 import utc from "dayjs/plugin/utc"
 import { FormikTouched, useFormik } from "formik"
-import { AutoStart } from "pages/WorkspacesPage/schedule"
-import { AutoStop } from "pages/WorkspacesPage/ttl"
+import { AutoStart } from "pages/WorkspaceSchedulePage/schedule"
+import { AutoStop } from "pages/WorkspaceSchedulePage/ttl"
 import { FC } from "react"
 import * as Yup from "yup"
 import { getFormHelpersWithError } from "../../util/formUtils"
